feat: カテゴリ管理機能を新たに実装し、カテゴリの追加、移動、表示を可能にする
